+++ pkg/ratelimit/ratelimiter.go
@@ -1,4 +1,4 @@
-package gateway
+package ratelimit
 
 import (
 	"errors"
