+++ pkg/ratelimit/ratelimiter_test.go
@@ -1,4 +1,4 @@
-package gateway
+package ratelimit
 
 import (
 	"testing"
